requirements-dev.txt
pytest
pyyaml
flake8